feat: move space page components up a level
