+++ apps/web/partials/editor/editor.tsx
@@ -13,6 +13,7 @@ import * as React from 'react';
 
 import { useEntityPageStore } from '~/core/hooks/use-entity-page-store';
 import { useHydrated } from '~/core/hooks/use-hydrated';
+import { useUserIsEditing } from '~/core/hooks/use-user-is-editing';
 
 import { SquareButton } from '~/design-system/button';
 import { Spacer } from '~/design-system/spacer';
@@ -25,7 +26,6 @@ import { ParagraphNode } from './paragraph-node';
 import { TableNode } from './table-node';
 
 interface Props {
-  editable?: boolean;
   placeholder?: React.ReactNode;
   shouldHandleOwnSpacing?: boolean;
 }
@@ -70,12 +70,9 @@ export const tiptapExtensions = [
   }),
 ];
 
-export const Editor = React.memo(function Editor({
-  placeholder = null,
-  shouldHandleOwnSpacing,
-  editable = true,
-}: Props) {
+export const Editor = React.memo(function Editor({ placeholder = null, shouldHandleOwnSpacing }: Props) {
   const { editorJson, spaceId, updateEditorBlocks, blockIds } = useEntityPageStore();
+  const editable = useUserIsEditing(spaceId);
 
   const editor = useEditor({
     extensions: [...tiptapExtensions, createIdExtension(spaceId)],
