+++ apps/web/partials/entity-page/toggle-entity-page.tsx
@@ -1,0 +1,31 @@
+'use client';
+
+import { useUserIsEditing } from '~/core/hooks/use-user-is-editing';
+import { Space, Triple } from '~/core/types';
+
+import { EditableEntityPage } from './editable-entity-page';
+import { ReadableEntityPage } from './readable-entity-page';
+
+interface Props {
+  triples: Triple[];
+  id: string;
+  name: string;
+  description: string | null;
+  spaceId: string;
+  serverAvatarUrl: string | null;
+  serverCoverUrl: string | null;
+
+  // For the page editor
+  blockTriples: Triple[];
+  blockIdsTriple: Triple | null;
+
+  space: Space | null;
+}
+
+export function ToggleEntityPage(props: Props) {
+  const renderEditablePage = useUserIsEditing(props.spaceId);
+
+  const Page = renderEditablePage ? EditableEntityPage : ReadableEntityPage;
+
+  return <Page {...props} />;
+}
